+++ datagrowth/configuration/__init__.py
@@ -1,6 +1,5 @@
-from datagrowth.settings import (DATAGROWTH_DEFAULT_CONFIGURATION,
-                                 DATAGROWTH_DEFAULT_CONFIGURATION as DEFAULT_CONFIGURATION)
 from datagrowth.configuration.types import (ConfigurationType, ConfigurationProperty, ConfigurationNotFoundError,
                                             create_config, register_defaults)
 from datagrowth.configuration.fields import ConfigurationField, ConfigurationFormField
 from datagrowth.configuration.serializers import load_config, DecodeConfigAction, get_standardized_configuration
+from datagrowth.configuration.defaults import DATAGROWTH_DEFAULT_CONFIGURATION, DEFAULT_CONFIGURATION